% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdaterAddin.R
\name{CacheTrue}
\alias{CacheTrue}
\title{Inserts cache option into copy of report.}
\usage{
CacheTrue(contents)
}
\arguments{
\item{contents}{Character vector, content of report's copy}
}
\value{
Caracter vector, updated content.
}
\description{
Inserts cache option into copy of report.
}
