% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdaterAddin.R
\name{Compare}
\alias{Compare}
\title{Calls for comparing python script.}
\usage{
Compare(echo.md.path, fair.id, name, fair)
}
\arguments{
\item{echo.md.path}{Character vector, path to document with .md extension that was knitted with echo option}

\item{fair.id}{Character vector, fair copy Gdoc id}

\item{name}{Character vector, the name of current .rmd document}

\item{fair}{Character vector, path to downloaded from Gdrive fair copy with .odt extension}
}
\value{
Character vector, python script's answer
}
\description{
Calls for comparing python script.
}
