% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdaterAddin.R
\name{CopyAndCompare}
\alias{CopyAndCompare}
\title{Creates copy of current report with echo option, knits it to .md file, calls comparation function.}
\usage{
CopyAndCompare(echo.true.report, fair.id, name)
}
\arguments{
\item{echo.true.report}{Character vector, with current report content copy with echo option}

\item{fair.id}{Character vector, fair copy Gdoc id}

\item{name}{Character vector, the name of current .rmd document}
}
\value{
Character vector, function's answer
}
\description{
Creates copy of current report with echo option, knits it to .md file, calls comparation function.
}
