% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdaterAddin.R
\name{Echo}
\alias{Echo}
\title{Creates copy of current report's content with echo option.}
\usage{
Echo(content)
}
\arguments{
\item{content}{Character vector, current report content}
}
\value{
Character vector, new content
}
\description{
Creates copy of current report's content with echo option.
}
