% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdaterAddin.R
\name{ExtractName}
\alias{ExtractName}
\title{Extracts name of current report.}
\usage{
ExtractName(path)
}
\arguments{
\item{path}{Character vector, current report path}
}
\value{
Character vector, extracted report name
}
\description{
Extracts name of current report.
}
