% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdaterAddin.R
\name{ExtractTitle}
\alias{ExtractTitle}
\title{Extracts title from YAML information in current report.}
\usage{
ExtractTitle(content)
}
\arguments{
\item{content}{Character vector, current report content}
}
\value{
Character vector, extracted title or
        NULL if failed
}
\description{
Extracts title from YAML information in current report.
}
