% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdaterAddin.R
\name{Find}
\alias{Find}
\title{Finds subvector.}
\usage{
Find(pattern, original, comparator = function(a, b) {     return(a == b) })
}
\arguments{
\item{pattern}{Any type vector - subvector to look for}

\item{original}{Any type vector to search in}

\item{comparator}{Function that compares two elements}
}
\value{
Number Vector with start indexes of entries of subvector, if pattern was founded
        Otherwise, zero length vector.
}
\description{
Finds subvector.
}
\examples{
Find(c(1, 2), c(4, 1, 2, 7, 1, 3, 2, 2, 1, 2))
Find(c("the", "Force"), c("May", "the", "Force", "be", "with", "you"))
Find(c("^motherf.?cker$"), c("English", "motherfucker", "do", "you", "speak", "it"), function(a,b){return(grepl(a,b))})
}
