% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Interface.R
\docType{data}
\name{interface}
\alias{interface}
\title{Shiny gadget interface.}
\format{An object of class \code{shiny.tag.list} (inherits from \code{list}) of length 3.}
\usage{
interface
}
\description{
Creates 2 tabs and 4 rows of buttons and 1 text field.
}
\details{
There are 2 tabs: one for buttons and text output and one for html output for diff table.
}
\keyword{datasets}
