% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdaterAddin.R
\name{PerformRefactor}
\alias{PerformRefactor}
\title{Finds patternt in content and replace it.}
\usage{
PerformRefactor(contents, from, to, useWordBoundaries = FALSE)
}
\arguments{
\item{contents}{Character vector, content to search in}

\item{from}{Character vector, regular expression, that will be replaced}

\item{to}{Character vector, that will be replacement}
}
\value{
List with new content and number of changes.
}
\description{
Finds patternt in content and replace it.
}
