% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdaterAddin.R
\name{Reupload}
\alias{Reupload}
\title{Reupload draft and fair copies on Gdrive with using existing links}
\usage{
Reupload(draft.id, fair.id, odt.report)
}
\arguments{
\item{draft.id}{Character vector, draft copy Gdoc id}

\item{fair.id}{Character vector, fair copy Gdoc id}

\item{odt.report}{Character vector, path to knitted report}
}
\value{
-
}
\description{
Reupload draft and fair copies on Gdrive with using existing links
}
