% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RMDupdaterAddin.R
\docType{package}
\name{RMDupdaterAddin-package}
\alias{RMDupdaterAddin}
\alias{RMDupdaterAddin-package}
\title{RMDupdaterAddin: What the Package Does (Title Case)}
\description{
More about what it does (maybe more than one line)
Use four spaces when indenting paragraphs within the Description.
}
\details{
The only function you're need from RMDupdaterAddin is [RMDupdaterAddin()].
}
\keyword{internal}
