% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdaterAddin.R
\name{RMDupdaterAddin}
\alias{RMDupdaterAddin}
\title{Shiny gadget.}
\usage{
RMDupdaterAddin()
}
\description{
Runs the shiny gadget with two tabs in Viewer section of RStudio.
}
\details{
RMDupdaterAddin was made for synchronization of .rmd file and human readable version on Gdrive.
It uses python script "RMD_updater" for comparing and algorythms based on regular expressions for
highlighting changes in .rmd document.
}
