% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdaterAddin.R
\name{SetCache}
\alias{SetCache}
\title{Inserts cache option into copy of report.}
\usage{
SetCache(contents, name)
}
\arguments{
\item{contents}{Character vector, content of report's copy}

\item{name}{String, name of rem file}
}
\value{
Caracter vector, updated content.
}
\description{
Inserts cache option into copy of report.
}
