% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdaterAddin.R
\name{SetOptions}
\alias{SetOptions}
\title{Creates copy of current report's content with echo and cache options.}
\usage{
SetOptions(content, name)
}
\arguments{
\item{content}{Character vector, current report content}

\item{name}{String, name of rmd file}
}
\value{
Character vector, new content
}
\description{
Creates copy of current report's content with echo and cache options.
}
