% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdaterAddin.R
\name{Update}
\alias{Update}
\title{Updates draft copy on Gdrive.}
\usage{
Update(session, draft.id, odt.report)
}
\arguments{
\item{draft.id}{Character vector, draft copy Gdoc id}

\item{odt.report}{Character vector, path to knitted report}
}
\value{
-
}
\description{
Updates draft copy on Gdrive.
}
