% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdaterAddin.R
\name{Upload}
\alias{Upload}
\title{Uploads knitted report in two copies on Gdrive and write Gdoc ids to sync_report.sh.}
\usage{
Upload(odt.report, report.name, report.name.draft, sync.path)
}
\arguments{
\item{odt.report}{Character vector, knitted report path.}

\item{report.name}{Character vector, fair copy name on Gdrive}

\item{report.name.draft}{Character vector, draft copy name on Gdrive}

\item{sync.path}{Character vector, path to sync_report.sh file}
}
\value{
-
}
\description{
Uploads knitted report in two copies on Gdrive and write Gdoc ids to sync_report.sh.
}
