% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UpdaterAddin.R
\name{Find}
\alias{Find}
\title{Finds subvector.}
\usage{
Find(pattern, original, comparator = function(a, b) {     return(a == b) })
}
\arguments{
\item{pattern}{Any type vector - subvector to look for}

\item{original}{Any type vector to search in}

\item{comparator}{Function that compares two elements}
}
\value{
Number Vector with start indexes of entries of subvector, if pattern was founded
        Otherwise, zero length vector.
}
\description{
Finds subvector.
}
